Inserindo confirmacao de cancelamento

+++ src/app/pages/excluir-livros/excluir-livros.component.ts
@@ -15,16 +15,37 @@ export class ExcluirLivrosComponent {
 
   livroId = 0;
   ngOnInit() {
-    this.livroId = Number(this.route.snapshot.paramMap.get('id'));
-    this.livrosService.excluirLivro(this.livroId).subscribe(result => {
-        Swal.fire(
-          'Tudo certo!',
-          'Cadastro excluído com sucesso!',
-          'success'
-        );
-        this.router.navigate(['/livros']);
-    }, error => {
-        console.error(error);
-    })
+    Swal.fire({
+      title: 'Você tem certeza?',
+      text: "Não será possíverl reverter essa ação!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Sim, excluir!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+
+        this.livroId = Number(this.route.snapshot.paramMap.get('id'));
+        this.livrosService.excluirLivro(this.livroId).subscribe(result => {
+            Swal.fire(
+            'Tudo certo!',
+            'Cadastro excluído com sucesso!',
+            'success'
+            );
+            this.router.navigate(['/livros']);
+        }, error => {
+            Swal.fire({
+            icon: 'error',
+            title: 'Oops...',
+            text: 'Algo deu errado! Verifique o erro no console.'
+            })
+            console.error(error);
+        })
+
+      }
+    this.router.navigate(['/livros']);
+    });
+
   }
 }
